Extract Button subcomponents to make linter happy and clean up Button file
